extracting non res historical and column names
