+++ jobs/estimate_job_counts.py
@@ -8,6 +8,24 @@ from pyspark.ml.evaluation import RegressionEvaluator
 from pyspark.sql import Window
 
 from utils import utils
+from utils.estimate_job_count.column_names import (
+    LOCATION_ID,
+    SERVICES_OFFERED,
+    PEOPLE_DIRECTLY_EMPLOYED,
+    NUMBER_OF_BEDS,
+    SNAPSHOT_DATE,
+    JOB_COUNT,
+    JOB_COUNT_SOURCE,
+    LOCAL_AUTHORITY,
+    REGISTRATION_STATUS,
+    ESTIMATE_JOB_COUNT,
+    ESTIMATE_JOB_COUNT_SOURCE,
+    PRIMARY_SERVICE_TYPE,
+    LAST_KNOWN_JOB_COUNT,
+)
+from utils.estimate_job_count.models.model_non_res_historical import (
+    model_non_res_historical,
+)
 
 from utils.prepare_locations_utils.job_calculator.job_calculator import (
     update_dataframe_with_identifying_rule,
@@ -20,21 +38,6 @@ NONE_NURSING_HOME_IDENTIFIER = "Care home without nursing"
 NONE_RESIDENTIAL_IDENTIFIER = "non-residential"
 
 # Column names
-LOCATION_ID = "locationid"
-LAST_KNOWN_JOB_COUNT = "last_known_job_count"
-ESTIMATE_JOB_COUNT = "estimate_job_count"
-ESTIMATE_JOB_COUNT_SOURCE = ESTIMATE_JOB_COUNT + "_source"
-PRIMARY_SERVICE_TYPE = "primary_service_type"
-PEOPLE_DIRECTLY_EMPLOYED = "people_directly_employed"
-NUMBER_OF_BEDS = "number_of_beds"
-REGISTRATION_STATUS = "registration_status"
-LOCATION_TYPE = "location_type"
-LOCAL_AUTHORITY = "local_authority"
-SERVICES_OFFERED = "services_offered"
-JOB_COUNT = "job_count"
-JOB_COUNT_SOURCE = "job_count_source"
-ASCWDS_IMPORT_DATE = "ascwds_workplace_import_date"
-SNAPSHOT_DATE = "snapshot_date"
 
 
 def main(
@@ -220,31 +223,6 @@ def populate_last_known_job_count(df):
     df = df.withColumn(LAST_KNOWN_JOB_COUNT, F.col("previous.job_count"))
     df = df.select(
         [f"current.{col_name}" for col_name in column_names] + [LAST_KNOWN_JOB_COUNT]
-    )
-
-    return df
-
-
-def model_non_res_historical(df):
-    """
-    Non-res : Historical :  : 2021 jobs = Last known value *1.03
-    """
-    # TODO: remove magic number 1.03
-    df = df.withColumn(
-        ESTIMATE_JOB_COUNT,
-        F.when(
-            (
-                F.col(ESTIMATE_JOB_COUNT).isNull()
-                & (F.col(PRIMARY_SERVICE_TYPE) == "non-residential")
-                & F.col(LAST_KNOWN_JOB_COUNT).isNotNull()
-            ),
-            F.col(LAST_KNOWN_JOB_COUNT) * 1.03,
-        ).otherwise(F.col(ESTIMATE_JOB_COUNT)),
-    )
-    df = update_dataframe_with_identifying_rule(
-        df,
-        "model_non_res_ascwds_projected_forward",
-        ESTIMATE_JOB_COUNT,
     )
 
     return df
