+++ utils/estimate_job_count/column_names.py
@@ -1,0 +1,15 @@
+LOCATION_ID = "locationid"
+LAST_KNOWN_JOB_COUNT = "last_known_job_count"
+ESTIMATE_JOB_COUNT = "estimate_job_count"
+ESTIMATE_JOB_COUNT_SOURCE = ESTIMATE_JOB_COUNT + "_source"
+PRIMARY_SERVICE_TYPE = "primary_service_type"
+PEOPLE_DIRECTLY_EMPLOYED = "people_directly_employed"
+NUMBER_OF_BEDS = "number_of_beds"
+REGISTRATION_STATUS = "registration_status"
+LOCATION_TYPE = "location_type"
+LOCAL_AUTHORITY = "local_authority"
+SERVICES_OFFERED = "services_offered"
+JOB_COUNT = "job_count"
+JOB_COUNT_SOURCE = "job_count_source"
+ASCWDS_IMPORT_DATE = "ascwds_workplace_import_date"
+SNAPSHOT_DATE = "snapshot_date"
